真·最终演示版

+++ src/com/stupidcoder/cc/lex/core/ICharPredicate.java
@@ -2,7 +2,7 @@ package com.stupidcoder.cc.lex.core;
 
 @FunctionalInterface
 public interface ICharPredicate {
-    boolean accept(byte input);
+    boolean accept(int input);
 
     static ICharPredicate single(int ch) {
         return c -> c == ch;
